refactor(backend-java): 外部連携を infra/ パッケージに集約
cognito / s3 / sqs / dynamo を controller/service と同じトップ階層に
並べていたのを、outbound adapter として infra/ 配下にまとめる。
分類の軸(技術レイヤ × 外部連携)が混在していたのを解消し、クリーン
アーキの「外部システム連携=インフラ層」を素直に表現する。

- com.normanblog.frestyle.{cognito,s3,sqs,dynamo}
  → com.normanblog.frestyle.infra.{cognito,s3,sqs,dynamo}
- ポート interface(ReportEnqueuer 等)+ 実装 + stub を各 infra サブパッケージに同居
- ロジック変更なし(package 宣言 + import のみ)。go build 相当の ./gradlew build 緑
- README のパッケージ構成図を更新(infra の役割 = 依存性逆転を明記)
