+++ backend-java/src/main/java/com/normanblog/frestyle/infra/s3/ProfileImagePresigner.java
@@ -1,4 +1,4 @@
-package com.normanblog.frestyle.s3;
+package com.normanblog.frestyle.infra.s3;
 
 import com.normanblog.frestyle.dto.ProfileImageUploadUrl;
