+++ backend-java/src/main/java/com/normanblog/frestyle/infra/sqs/ReportEnqueuer.java
@@ -1,4 +1,4 @@
-package com.normanblog.frestyle.sqs;
+package com.normanblog.frestyle.infra.sqs;
 
 /** 学習レポート生成ジョブを非同期キュー(SQS)に投入する境界。 */
 public interface ReportEnqueuer {
